refactor(services): moves all services from 'services' folder into 'service'
