+++ role/views.py
@@ -7,7 +7,7 @@ from django.views.decorators.csrf import csrf_protect, csrf_exempt
 from .models import Role
 from .normalizers import role_normalizer, roles_normalizer
 from .forms import RoleForm
-from services.send_response import send_json_response
+from service.send_response import send_json_response
 
 
 # @csrf_protect
